+++ infra/settings_app_infra.py
@@ -15,8 +15,6 @@
 import textwrap
 from typing import List, Tuple
 
-import datarobot as dr
-import pulumi
 import pulumi_datarobot as datarobot
 
 from forecastic.i18n import LanguageCode, LocaleSettings
@@ -38,17 +36,6 @@ app_source_args = ApplicationSourceArgs(
 
 app_resource_name: str = f"Forecast Assistant Application [{project_name}]"
 application_locale = LocaleSettings().app_locale
-
-
-def ensure_app_settings(app_id: str) -> None:
-    try:
-        dr.client.get_client().patch(
-            f"customApplications/{app_id}/",
-            json={"allowAutoStopping": True},
-            timeout=60,
-        )
-    except Exception:
-        pulumi.warn("Could not enable autostopping for the Application")
 
 
 def _prep_metadata_yaml(